fastapi/requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cos-python-sdk-v5==1.9.15
crcmod==1.7
cryptography==3.4.8
dicttoxml==1.7.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.961
mypy-extensions==0.4.3
optionaldict==0.1.2
orjson==3.7.7
passlib==1.7.4
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
requests-pkcs12==1.14
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
tencentcloud-sdk-python==3.0.618
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
watchfiles==0.15.0
watchgod==0.8.2
websockets==10.3
wechatpy==2.0.0a11
xmltodict==0.13.0
